Issue #2807: extracted FxaReceivedTab to its own file
This does not yet compile. Fixed in the following commit
